feat(codegen): generation code output
